+++ integration-tests/gatsby-source-wordpress/docker/wordpress/install-wp-graphql-plugins.sh
@@ -15,10 +15,10 @@ git clone --depth 1 -b ${WPGATSBY_VERSION} https://github.com/gatsbyjs/wp-gatsby
 # WP GraphQL ACF from GitHub release
 install-plugin wp-graphql-acf wp-graphql/wp-graphql-acf ${WPGRAPHQL_ACF_VERSION} && \
 
-# wp-graphql-gutenberg from Github release
+# wp-graphql-gutenberg from GitHub release
 install-plugin wp-graphql-gutenberg pristas-peter/wp-graphql-gutenberg ${WPGRAPHQL_GUTENBERG_VERSION} && \
 
-# CPTUI from Github release
+# CPTUI from GitHub release
 install-plugin custom-post-type-ui WebDevStudios/custom-post-type-ui ${CPT_UI_VERSION} && \
 
 # WP GraphQL CPTUI from GitHub release