+++ packages/gatsby-source-wordpress/src/steps/create-schema-customization/build-types.js
@@ -130,7 +130,7 @@ const objectType = typeBuilderApi => {
   // if all child fields are excluded, this type shouldn't exist.
   // check null first, otherwise cause:
   // TypeError: Cannot convert undefined or null to object at Function.keys (<anonymous>)
-  // Also cause wordpress blog site build failure in createSchemaCustomization step
+  // Also cause WordPress blog site build failure in createSchemaCustomization step
   if (!transformedFields || !Object.keys(transformedFields).length) {
     return false
   }