+++ packages/gatsby/src/redux/plugin-runner.ts
@@ -1,7 +1,8 @@
 import { Span } from "opentracing"
-import { emitter } from "./index"
+import { emitter, store } from "./index"
 import apiRunnerNode from "../utils/api-runner-node"
 import { ActivityTracker } from "../../"
+import { ICreateNodeAction } from "./types"
 
 type Plugin = any // TODO
 
@@ -41,12 +42,35 @@ interface ICreatePageAction {
 }
 
 export const startPluginRunner = (): void => {
-  emitter.on(`CREATE_PAGE`, (action: ICreatePageAction) => {
-    const page = action.payload
-    apiRunnerNode(
-      `onCreatePage`,
-      { page, traceId: action.traceId, parentSpan: action.parentSpan },
-      { pluginSource: action.plugin.name, activity: action.activity }
-    )
-  })
+  const plugins = store.getState().flattenedPlugins
+  const pluginsImplementingOnCreatePage = plugins.filter(plugin =>
+    plugin.nodeAPIs.includes(`onCreatePage`)
+  )
+  const pluginsImplementingOnCreateNode = plugins.filter(plugin =>
+    plugin.nodeAPIs.includes(`onCreateNode`)
+  )
+  if (pluginsImplementingOnCreatePage.length > 0) {
+    emitter.on(`CREATE_PAGE`, (action: ICreatePageAction) => {
+      const page = action.payload
+      apiRunnerNode(
+        `onCreatePage`,
+        { page, traceId: action.traceId, parentSpan: action.parentSpan },
+        { pluginSource: action.plugin.name, activity: action.activity }
+      )
+    })
+  }
+
+  // We make page nodes outside of the normal action for speed so we manually
+  // call onCreateNode here for SitePage nodes.
+  if (pluginsImplementingOnCreateNode.length > 0) {
+    emitter.on(`CREATE_NODE`, (action: ICreateNodeAction) => {
+      const node = action.payload
+      if (node.internal.type === `SitePage`) {
+        apiRunnerNode(`onCreateNode`, {
+          node,
+          traceTags: { nodeId: node.id, nodeType: node.internal.type },
+        })
+      }
+    })
+  }
 }