+++ packages/gatsby/src/redux/types.ts
@@ -195,6 +195,7 @@ export type GatsbyNodeAPI =
   | "onPostBoostrap"
   | "onCreateWebpackConfig"
   | "onCreatePage"
+  | "onCreateNode"
   | "sourceNodes"
   | "createPagesStatefully"
   | "createPages"