+++ packages/gatsby/src/services/initialize.ts
@@ -104,10 +104,6 @@ export async function initialize({
     args.setStore(store)
   }
 
-  // Start plugin runner which listens to the store
-  // and invokes Gatsby API based on actions.
-  startPluginRunner()
-
   const directory = slash(args.directory)
 
   const program = {
@@ -262,6 +258,10 @@ export async function initialize({
     plugins: pluginsStr,
   })
 
+  // Start plugin runner which listens to the store
+  // and invokes Gatsby API based on actions.
+  startPluginRunner()
+
   // onPreInit
   activity = reporter.activityTimer(`onPreInit`, {
     parentSpan,
@@ -415,7 +415,10 @@ export async function initialize({
         ]
 
         if (process.env.GATSBY_EXPERIMENTAL_PRESERVE_FILE_DOWNLOAD_CACHE) {
-          // Add gatsby-source-filesystem
+          // Stop the caches directory from being deleted, add all sub directories,
+          // but remove gatsby-source-filesystem
+          deleteGlobs.push(`!${cacheDirectory}/caches`)
+          deleteGlobs.push(`${cacheDirectory}/caches/*`)
           deleteGlobs.push(`!${cacheDirectory}/caches/gatsby-source-filesystem`)
         }