+++ packages/gatsby-recipes/CHANGELOG.md
@@ -3,6 +3,14 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+# [0.17.0-next.0](https://github.com/gatsbyjs/gatsby/compare/gatsby-recipes@0.16.0-next.1...gatsby-recipes@0.17.0-next.0) (2021-05-06)
+
+**Note:** Version bump only for package gatsby-recipes
+
+# [0.16.0-next.1](https://github.com/gatsbyjs/gatsby/compare/gatsby-recipes@0.16.0-next.0...gatsby-recipes@0.16.0-next.1) (2021-05-05)
+
+**Note:** Version bump only for package gatsby-recipes
+
 # [0.16.0-next.0](https://github.com/gatsbyjs/gatsby/compare/gatsby-recipes@0.15.0-next.3...gatsby-recipes@0.16.0-next.0) (2021-04-23)
 
 **Note:** Version bump only for package gatsby-recipes